requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
amqp==5.1.1
    # via
    #   -r requirements/base.txt
    #   kombu
asgiref==3.5.2
    # via
    #   -r requirements/base.txt
    #   django
attrs==22.1.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
    #   pytest
billiard==3.6.4.0
    # via
    #   -r requirements/base.txt
    #   celery
celery==5.2.7
    # via -r requirements/base.txt
certifi==2022.9.24
    # via
    #   -r requirements/base.txt
    #   requests
charset-normalizer==2.1.1
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.3
    # via
    #   -r requirements/base.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   code-annotations
click-didyoumean==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.2.0
    # via
    #   -r requirements/base.txt
    #   celery
code-annotations==1.3.0
    # via -r requirements/test.in
coverage[toml]==6.5.0
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   openedx-events
edx-opaque-keys[django]==2.3.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
fastavro==1.6.1
    # via
    #   -r requirements/base.txt
    #   openedx-events
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==1.1.1
    # via pytest
jinja2==3.1.2
    # via code-annotations
kombu==5.2.4
    # via
    #   -r requirements/base.txt
    #   celery
markupsafe==2.1.1
    # via jinja2
    # via -r requirements/base.txt
packaging==21.3
    # via pytest
pbr==5.10.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.31
    # via
    #   -r requirements/base.txt
    #   click-repl
py==1.11.0
    # via pytest
pymongo==3.12.3
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pyparsing==3.0.9
    # via packaging
pytest==7.1.3
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-slugify==6.1.2
    # via code-annotations
pytz==2022.4
    # via
    #   -r requirements/base.txt
    #   celery
    #   django
pyyaml==6.0
    # via code-annotations
requests==2.28.1
    # via -r requirements/base.txt
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   click-repl
sqlparse==0.4.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==4.0.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
urllib3==1.26.12
    # via
    #   -r requirements/base.txt
    #   requests
vine==5.0.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit
